+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 | Version      |    Date     | Notes |
 |--------------|:-----------:|-----------------------------------------------------|
+| 3.1.         |   4/24/24   | Updated evaluation library to version 2.0.5 and added ARIA 1.3 option
 | 3.0.4        |   3/4/24   | Updated evaluation library to version 2.0.4
 | 3.0.2        |   2/29/24   | Updated evaluation library to version 2.0.3
 | 3.0.1        |   2/7/24    | Updated evaluation library
@@ -24,8 +25,10 @@
 
 =======
 
-### 3.05
+### 3.1
 * Fixed bug with element result details showing another result for the same element
+* Updated evaluation library
+* Added option to use ARIA 1.3 requirements for validation
 
 ### 3.04
 * Evaluation library update fixes bug in widget rule for required parent roles
@@ -34,7 +37,7 @@
 * Update includes fixed bug in color contrast calculation
 
 ### 3.01
-* Updated evaluation library to versino 2.0.2 with improved support of row role in table, grid and treegrid; and color contrast rule
+* Updated evaluation library to version 2.0.2 with improved support of row role in table, grid and treegrid; and color contrast rule
 
 ### 3.0
 * Updated video and audio rules